requirements.txt
RPi.GPIO>=0.7.0
Adafruit-Blinka>=6.5.0
adafruit-circuitpython-neopixel>=6.0.3
python-escpos==2.2.0
django>=3.2.3
